browser-use-tests/requirements.txt
playwright>=1.42.0
openai>=1.0.0
python-dotenv>=1.0.0
pytest>=7.4.0
pytest-asyncio>=0.21.1
pydantic>=2.4.0
loguru>=0.7.0